Refactor file system logic of StandaloneTrustedVaultBackend.
Refactor file system logic (reading, writing, migration) of
StandaloneTrustedVaultBackend into StandaloneTrustedVaultStorage.

This reduces complexity and size of StandaloneTrustedVaultBackend.

Note: This change introduces a small behavioural change. Previously,
upon migration of the state on disk, delegates' NotifyStateChanged()
method was called for every intermediary migration step. This no longer
is the case, delegates aren't notified at all during migration. As far
as I could tell, delegates are only used in tests and in
StandaloneTrustedVaultClient::NotifyBackendStateChanged(), which only
notifies debug observers. This particular behaviour around the migration
doesn't seem to be relied upon currently.

Also note: StandaloneTrustedVaultBackendTest was only adapted in a
minimal way to work with StandaloneTrustedVaultStorage. This was done on
purpose to verify that this change really only is a refactoring and
doesn't introduce behavioural changes. StandaloneTrustedVaultBackendTest
will be refactored in a follow-up change to no longer rely on disk
access but instead work with a fake implementation of
StandaloneTrustedVaultStorage.

BUG: 398160323
Change-Id: I9bda326fae0b2db2cd088cf1196fa311e4a79b27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6275946
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Thomas Thrainer <thomasth@google.com>
Cr-Commit-Position: refs/heads/main@{#1424369}