GetSourceImageInternal: Use kDontChangeAlpha by default for canvas
The function OffscreenCanvas::GetSourceImageForCanvas indicates a
default value of kPremultiplyAlpha. This was never used (the parameter
is always specified by the callers). Remove it.

The function HTMLCanvasElement::GetSourceImageForCanvasInternal also
indicates a default value of kPremultiplyAlpha. This default value was
used by HTMLCanvasElement::NotifyListenersCanvasChanged.

In HTMLCanvasElement::NotifyListenersCanvasChanged, the image is passed
to StaticBitmapImageToVideoFrameCopier::Convert. Inside that function:
- There exists a fast path for non-premultiplied (opaque or
  unpremultiplied) alpha, which will call directly into
  media::CreateFromSkImage. We may now start hitting this fast path.
- Otherwise, the result is passed to ReadARGBPixelsSync, which will
  convert the content to unpremultiplied alpha if it is not already.

Bug: 40277213
Change-Id: I1a40cc4ef52f7688d67210fcde8cb58d9111b0b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298234
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424370}