arrayify: Prefer `constexpr`
This CL is intended to be trivial. It marks the return from `std::to_array` as
`constexpr`. This CL is a split of a naive textual search-and-replace under:

components/drive

This CL was uploaded by git cl split.

Change-Id: Ied763ce6c959dbc58fefc200022bf4533301f340
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299470
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424375}