[DevTools] Store automatic file systems in `devtools.file_system_paths`.
The previous approach with storing the automatic file systems separately
from the other file systems (in `devtools.automatic_file_systems` prefs)
caused some issues in maintaining a consistent state between these two
separate lists. With this CL we now store all file systems, including
the automatic ones, in the `devtools.file_system_paths` prefs; for the
automatic file systems we use their UUID as the `type` (and thereby
forbit using an UUID as type for other file systems).

This makes it easier to reason about the state of the file systems
and removes one layer of complexity. The obsolete pref key will be
removed in a separate CL.

It also adds some extra validation, in particular, we now check that
the `file_system_path` that is passed when connecting an automatic
file system actually refers to a local folder (in addition to it being
a valid absolute path), and we check that `file_system_uuid` is a valid
UUID.

The CL also introduces error reporting via `FileSystemAdded` event to
the automatic file system handling, similar to what we already do for
the manual file systems.

Bug: 398973989, 395562934
Doc: http://go/chrome-devtools:automatic-workspace-folders-design
Change-Id: I1a6c7d7e117aa7d361749bd245b51e1b7b7734f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298193
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424396}