Update dialog layout and remove scrollbars
This CL replaces the previous call to `SetUseDefaultFillLayout` with
explicitly providing a FlexLayout in the Parent access, which view
gives ur more flexibility to configure the margins and the children
views arrangement.

Invoking `EnableSizingFromWebContents` to fill in the view with the web
contents results in removing the scroll bars.

See crbug.com/394839768#comment4 for a screen recording.

Bug: 394839768, 385358451
Change-Id: I1f9c98418a05a9f638ccd962f9c2fac67864ecd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288670
Reviewed-by: Duong Dac <ddac@chromium.org>
Commit-Queue: Anthi Orfanou <anthie@google.com>
Cr-Commit-Position: refs/heads/main@{#1424404}