arrayify: Prefer `constexpr`
This CL is intended to be trivial. It marks the return from `std::to_array` as
`constexpr`. This CL is a split of a naive textual search-and-replace under:

net/spdy

This CL was uploaded by git cl split.

Change-Id: Ide306cc5aec68b4a3fffb8da2823697503ce3ad6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6301528
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424514}