Set SiteForCookies in URLDownloader::FetchPDFFile
ResourceRequest.site_for_cookies defaults to an empty SiteForCookies.
When UCC launches, when users opt to restrict 3P cookies a request
with an empty SFC will have its cookies blocked by 3PC blocking.

This CL sets the SiteForCookies to the site of the request URL. This
will send cookies when 3PC blocking is enabled. It will also send
SameSite cookies.

If this request is meant to be cross-site or we do not want to send
cookies in this context, the change should be ignored. In that case,
we may want to change `credentials_mode` to `omit` to indicate that
the request explicitly does not intend to include cookies.

Bug: 324879760
Change-Id: If62e6d49876471f506ee209c85a1122054533839
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299241
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/main@{#1424516}