[ios blink] Fix message IO test failures
Now that we have more tests passing fix the remaining IO based tests.
These tests were failing because of the way the MessagePumpIO is
allocated in tests. Instead of the complex dance of trying to find
the right sequneced task runner just look it up each time. Use
a thread checker which is dynamic to test that the call happens
on the right thread.

Bug: 40254930
Change-Id: I536ee720ea046fe630fb8e6ac8368a38949853bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298840
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424546}