Fix CHECK failure in variations seed fuzz test
This fixes the fuzz testing but dropping the randomized data that
is semantically invalid and replacing it with dummy value.

`base::Version` allows default constructor but as per comment on it
"The only thing you can legally do to a default constructed Version
object is assign to it.".

`CreateTrialsFromSeedTestCase::ClientFilterableState` represents the
version as a repeated integer, and we were using default constructed
value when it was empty. Instead we use a `0.0.0.0` placeholder to
only run the code on the valid data.

TESTED=run a test values from bug on the modified code and now it passes

Bug: 392716876
Change-Id: I2367c89b55bb3ef659250f21bc02b0ab353590fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6289449
Reviewed-by: James Lee <ljjlee@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424557}