[search aggregator] Change expectation of UpdateResults
This CL changes the `UpdateResults` logic such that it no longer returns
a boolean and instead handles the notifying logic previously in
`RequestCompleted`. Since we will be moving to use DataDecoder for
parsing instead of JSONReader (to be added in a follow up CL), we no
longer need an UpdateResults method that returns a boolean. Instead we
will switch to an async control flow.

Bug: 395938335
Change-Id: I75a6d657b85546861d8fa054a7e881ed7ae5e783
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6282127
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Ananya Seelam <ananyaseelam@google.com>
Cr-Commit-Position: refs/heads/main@{#1424576}