arrayify: Prefer `constexpr`
This CL is intended to be trivial. It marks the return from `std::to_array` as
`constexpr`. This CL is a split of a naive textual search-and-replace under:

sandbox/linux

This CL was uploaded by git cl split.

Change-Id: I9cb603f1788451ee4312bb7c5703b2e698b0348f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299474
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424578}