[iOS][MIM2] Use IdentityManager::OnEndBatchOfPrimaryAccountChanges()
Update the NTP particle to listen for
IdentityManager::OnEndBatchOfPrimaryAccountChanges() instead of
IdentityManager::OnPrimaryAccountChanged().

The new API has been added with crrev.com/c/6280786 and
crrev.com/c/6298116.

The goal is to make sure the particle disk is not updated during
the transition from being signed in with one identity to be signed
in with another identity.
The signed out particle won't be shown.

Bug: 397194707, 397194715
Change-Id: Id41c83cb2de6e51699b737b97505b4399e4f9031
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299056
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424588}