arrayify: Prefer `constexpr`
This CL is intended to be trivial. It marks the return from `std::to_array` as
`constexpr`. This CL is a split of a naive textual search-and-replace under:

net/dns

This CL was uploaded by git cl split.

Change-Id: I8c09fb0a501655383c927ac96cd7b7e5b3076200
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6301269
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424607}