Create an InstanceIdentityTokenGetter for Cloud hosts
This CL updates the Me2Me host to create an instance of the
InstanceIdentityTokenGetter and pass it to the Cloud heartbeart and
SessionAuthz classes. I'll follow up with Cls which retrieve a
token for each of these classes in a follow-up.

Bug: 388885661
Change-Id: I8f4964c2905e40d396f26fdd83c9091439e516e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299998
Auto-Submit: Joe Downing <joedow@chromium.org>
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Commit-Queue: Gary Kacmarcik <garykac@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424620}