Build Safe Browsing download protection on Android (4/6)
This is part 4 of 6 for building code for download protection on
Android. The overall strategy is to define a new buildflag,
SAFE_BROWSING_DOWNLOAD_PROTECTION, and use it to guard all
code related to Safe Browsing download protection, and then
add Android to the platforms on which it is true.

In this CL, code related to file analyzers and the file util service
is explicitly excluded from building on Android. These analyzers for
RAR, ZIP, 7Z, and DMG archive types do not apply to Android.

Bug: 397407934
Change-Id: I4d6a43061dcfeff50af17eff5dc3eb80becfe3a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6292867
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424665}