[A11y] IsKnownScreenReaderAppActive() + VoiceOver histogram + crash key
1. Implement IsKnownScreenReaderAppActive() for Mac/Win.
This cannot be based on AXMode::kScreenReader because that
an inaccurate heuristic. We need something that can be updated
based on whether an actual screen reader app is currently running.
2. Add VoiceOver histogram + crash key (these already exist for Windows
screen readers).

IsKnownScreenReaderAppActive() is needed to vet the upcoming
refreshed auto-disable heuristic -- it will report whenever we are about
to auto disable a11y (when we think a11y is no longer used), but a
screen reader is still active, which indicates a flaw in the heuristic
logic.

Auto-disable refresh:
https://docs.google.com/document/d/1DlJn8PPE4towT0W4VknBqB04-SzFAz6yZ0graPc_xS8

To do in follow-ups, implement IsKnownScreenReaderAppActive() for:
* Talkback on Android
* ChromeVox on ChromeOS
* Orca on Linux

Note that these cannot be based on

Bug: none
Change-Id: I3c16dfbe8a5dd1f41b0f854c3a744dc4783b5cc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299355
Reviewed-by: Jayson Adams <shrike@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424679}