[iOS] Replace the accessory's gradient view with a layer mask
This CL replaces the gradient view initially placed on the
accessory view with a layer mask placed on a new container
view added to contain the form suggestion view. The reason
for adding a new container view is that the form suggestion
view is a scroll view and the gradient mask should not move
along with the scroll view.

The reason why the gradient view is being removed is that it
works only as long as the background color of the content
view underneath it does not change, which won't be the case
if the background is made translucent.

Bug: 385172448
Change-Id: Ieedfddd5971ff1d1356ee2fdbbc895f2a28613ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299548
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1424686}