Fix glic pinning metrics
Previously in crrev.com/c/6292539 I incorrectly added a histogram to log
the glic pinning pref. It was supposed to be a user action, and also
we needed to update the impression logic to take into account the states
of the os and browser entry points.

Fixed: 398265399
Change-Id: Id53a60fd7d90fd5ef5c53ee13b3aecf2fe6289be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298411
Reviewed-by: Andrea Gomez <andreaxg@google.com>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424696}