Build Safe Browsing download protection on Android (6/6)
This is part 6 of 6 for building code for download protection on
Android. The overall strategy is to define a new buildflag,
SAFE_BROWSING_DOWNLOAD_PROTECTION, and use it to guard all
code related to Safe Browsing download protection, and then
add Android to the platforms on which it is true.

In this CL, the buildflag is defined to be true on Android. Source
files are moved to the appropriate sections in BUILD.gn files to start
building Safe Browsing download protection code on Android.

A temporary placeholder null implementation for
ServicesDelegateAndroid's instantiation of DownloadProtectionService
ensures that no download protection requests are actually sent yet for
Android. They will be further implemented as part of the
MaliciousApkDownloadCheck feature.

Bug: 397407934
Change-Id: Ifc69dacbe276848023fbaad126bbfc92359dcf62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6293364
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424739}