Integrate price drop with all tab resumption integration points
There are 3 paths to Tab Resumption data acquisition. I thought
the path with VisitedURLRankingService was the final state although
this was just an experiment (which isn't going ahead). This CL
integrates with all 3 for simplicity - recent local tab, recent
synced tab and VisitedURLRankingService. This will make it much
easier to write dogfood/testing instructions for QA as it will
ensure the price drop chip will show in all scenarios (where
it exists for the URL).

Bug: 398852977
Change-Id: I49b673989be7dde4059895796d3b0e0f93723f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299850
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1424742}