Implement accept callback for BNPL ToS dialog in BnplManager
This CL implements the accept callback for the BNPL ToS dialog in the
BnplManager. This acceptance callback needs to load risk data, and send
a CreateBnplPaymentInstrument request to the Payments server.

As part of this change, there were also a couple of code health items
taken care of:

- We should retrieve the app locale from `AutofillClient`, so this CL
passes in AutofillClient instead of PaymentsAutofillClient in the
BnplManager constructor.

- BnplManager::ShouldShowBnplSettingsToggle() was out of order in the
.h vs .cc files, so I moved the .cc file implementation.

- BnplManager::GetBnplSupportedCountries() was unused, so I removed it.

- I updated the type of the context token in BNPL requests from an
std::u16string to an std::string.

Bug: 356443046, b:378518488
Change-Id: I01e99708dcd63a9699b95f318f039fe0226e3215
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6283169
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424752}