Add net error code for self-signed certs on local network URLs
This adds (behind a new, default disabled feature flag) a new net error
code (ERR_CERT_SELF_SIGNED_LOCAL_NETWORK), which is returned for
certificate errors when the certificate is self-signed, and is being
used on a local network URL (either an RFC 1918 IP, or a URL that ends
in .local).

This is step one towards implementing go/betterselfsigned, for now this
only results in the regular SSL interstitial being triggered.

Bug: 394119724
Change-Id: I62850128c250358071cc0693526b23c2677647cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6226147
Commit-Queue: Carlos IL <carlosil@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424804}