Ensure that chip is always in the front
Page action that is in suggestion chip mode should be always in the
front and when the state change it should back to it initial position.

The current contract is that only one suggestion chip will be visible at
a time. The page action controller will give that guarantee.

Bug: 395118892
Change-Id: Ia40fe99ccfdf53fc88d6b81a02eaf5cb6565d3db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6242576
Commit-Queue: Foromo Daniel Soromou <koretadaniel@chromium.org>
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424817}