[ios blink] Disable crashpad initialization
Crashpad was crashing when initializing. I am not sure if it is
related to the entitlements or the fact that the process is loaded
as a framework, and another thread is resolving symbols. But for
now we just disable its initialization so we can have a build
running.

Bug: 40254930
Change-Id: Ib668911826f8a96aeb76318eab13b78dcd1a387b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6299290
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424822}