Consider mime type for extension file verification
The browser tries to harden extension verification by always verifying
JavaScript and HTML files, but the way it does that is incomplete: it
checks for the .js, .html, and .htm extensions, and misses several
others. Change the code to consider the mime type instead, similar to
the logic that's used when serving the extension files.

Bug: 398554830
Change-Id: I6d33c71c98590359bb96f725b3da90b2a5475bae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6294985
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Michael Maltsev (R&D) <michaelm@island.io>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424831}