Fix ResolveStopColorWithMissingParams logic
When the color stop have different colorspace with the neighbor, the
components is mixing with each other, which is clearly wrong.

This patch will convert the neighbor to the same colorspace with the one
is been resolved first.

Bug: 399046941
Change-Id: Icb6d8aff43c8d8301a274025dde3749e2e929d0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288130
Auto-Submit: Jason Leo <m.jason.liu@gmail.com>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424834}