sunfish: Rename `IsSunfishSessionAllowed`
The name "Sunfish-session" is generally confusing. After some recent
decoupling, `IsSunfishSessionAllowed` is now a simple check for whether
Sunfish-related UI or Scanner-related UI can be shown.

Rename the function to match, which should be much more understandable
in the places that it is used.

Bug: b:397784915
Change-Id: I6a6a210060e384a2ccff69691497524d7f504684
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6290669
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1424867}