sunfish: Rename methods called from `ShowDefaultActionButtons…`
`CaptureModeSession::ShowDefaultActionButtonsOrPerformSearch` is the
only caller of two `CaptureModeBehavior` methods:
`ShouldShowDefaultActionButtonsAfterRegionSelected` and
`OnRegionSelectedOrAdjusted`.

These methods are only called _after_
`CaptureModeSession::ShouldShowActionContainerWidget()` is checked,
which includes checks for `CanShowSunfishOrScannerUi()`, the selection
being an image region, and the behavior `CanShowActionButtons()`.

Document the checks that `ShouldShowActionContainerWidget` does in the
header file, and rename
`ShouldShowDefaultActionButtonsAfterRegionSelected` and
`OnRegionSelectedOrAdjusted` to be explicit that they are only used when
the action container should be shown.

Making this rename makes it clear that some checks are redundant, and
can be removed to simplify the logic around action buttons. This will be
done in a follow-up CL.

Bug: b:397296160
Change-Id: I6a6a2100964fd14124b8903913e73b83c770548a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6275812
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1424877}