arrayify: Prefer `constexpr`
This CL is intended to be trivial. It marks the return from `std::to_array` as
`constexpr`. This CL is a split of a naive textual search-and-replace under:

components/translate

This CL was uploaded by git cl split.

Change-Id: I608db52de669e200ffe12baacd55dce4211794b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298272
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424959}