extensions: Refactor ExtensionService GrantPermissionsAndEnableExtension
This is part of the effort to get CrxInstaller compiling on desktop
Android. GrantPermissionsAndEnableExtension's implementation is just
two lines, using PermissionUpdater and ExtensionRegistrar. Move the
impl to ExtensionRegistrar, then call it from both CrxInstaller and
ExtensionService (which we can't yet compile on desktop Android).

Bug: 396722906
Change-Id: I6c084e5698d9f9d7ae8ff4bc24f26d3b0d4721e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6302374
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424967}