[iOS][LO] Set missing top offset for post capture selection
For the screenshot based lens overlay the top offset is set
automatically, so the zero value was ignored. This is not the case for
LVF Post Capture, which needs the top offset to be set manually.

Also adjust the bottom offset a bit to reduce the padding when the full
image is selected.

Bug: None
Change-Id: I583835162c14cf7baf71118be5b7c938c30d0b27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6298150
Auto-Submit: Radu Nitescu <radunitescu@google.com>
Commit-Queue: Radu Nitescu <radunitescu@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425041}